Added puid project identifier argument to the initial deploy script.
Added ssr port to the config so ssr script uses that port to spawn a server.

Updated nginx config on initial deploy.

Removed start argument from git script. Start is the default behaviour.

Added custom nginx config support as it will be included in the main config of the host.